+++ src/core/libraries/libs.cpp
@@ -16,11 +16,11 @@
 #include "core/libraries/np_score/np_score.h"
 #include "core/libraries/np_trophy/np_trophy.h"
 #include "core/libraries/pad/pad.h"
+#include "core/libraries/save_data/savedata.h"
 #include "core/libraries/screenshot/screenshot.h"
 #include "core/libraries/system/commondialog.h"
 #include "core/libraries/system/msgdialog.h"
 #include "core/libraries/system/posix.h"
-#include "core/libraries/system/savedata.h"
 #include "core/libraries/system/savedatadialog.h"
 #include "core/libraries/system/sysmodule.h"
 #include "core/libraries/system/systemservice.h"
